Divide el controller de factura en 3, Factura, FacturaVenta y FacturaCompra.
